requirements.txt
fastapi
uvicorn
requests
python-dotenv
streamlit 
pydantic
sendgrid